refactor(velocity-ide): apply clippy --fix for 60 machine-applicable lints
Auto-fixed via cargo clippy --fix -p velocity-ide --all-targets. Categories cleared: manual_range_contains, bool_assert_comparison, len_zero, useless_vec, identity_op, needless_borrows_for_generic_args, unnecessary_cast, clone_on_copy, neg_multiply, bool_comparison, redundant_closure, manual_div_ceil, useless_conversion, manual_is_multiple_of. All MachineApplicable (behavior-preserving). 28 files, 59 insertions / 63 deletions. Byte-preserving: no BOM stripped (0 line-1 hunks), U+FFFD=0. Verified clippy -p velocity-ide EXIT=0 with 0 errors; 43 warnings remain (fixed in follow-up).

+++ velocity-ide/src/model/transformer_zero.rs
@@ -1466,7 +1466,7 @@ mod tests {
         assert_eq!(report.total_tokens(), n + 100);
         assert!((report.cache_hit_rate() - 0.5).abs() < 1e-9);
         assert!((report.us_per_token() - 1000.0).abs() < 1e-9);
-        assert_eq!(report.token_ids.len(), n as usize);
+        assert_eq!(report.token_ids.len(), n);
     }
 
     #[test]
